Improved the admin service, which tracks a logged in user's permission level
Added comments and picked more descriptive names. Made sure that the service always kept a separate current value for the model(you can think of this service as maintaining the model)

Emphasized that the Behavior Subject is a datastream which any subscribed components are listening to, so they can dynamically change their content as soon as the permission level changes.