First half of updates to DossierViewHelper
In flask backend, made methods that provide observables of REST service endpoints that (1)get all observations associated with a seal/dossier id, and (2) get the most recent observation of-that-seal/in-that-dossier

In dossier view helper(which keeps the data needed by the dossier view up to date), added Behavior subjects which really should have been Observables. Provided getters for those observables.