Changes to SqlObservation class, and the Seal/Dossier View
added better icons

made the height of the carousel fixed so it didn't shift the contents of the panel around.

Almost completed populating the panel with information. Only the Date Last Seen remaining.

Added an error handler class in class problems occurred between the flask-backend angular service and the flask rest api.

I was improperly representing some date objects as strings. Changed the SqlObservation class to reflect that Date, FirstSeenWeaner, and LastSeenPup were all dates, WeanDateRange was not a Date, but a number of days.