Got the creation of a request for a new user working again
Requests for a new user account are now created in the database. They are represented as parallel tuples in the Users and Observers entity sets, which have a value of 0 for their attribute "isVerifiedByAdmin". If the desired email already exists in the database for some other user, the operation fails.