Eliminated CORS error, Appropriate error JSON response returned to angular application from the Flask API.
Got rid of the annoying CORS error. It actually had nothing to do with CORS. When someone tried to log in as a user with a username that wasn't in the database, the code accessed elements of a database query searching for a match. Since the response was empty, checking the attributes(which did not exist) of the response caused an exception to be raised.

Also cleaned up the login-state component, the sign-in component, auth.service.ts, and the angular service flask-backend.service.ts