Fixed login after password change bug, various changes related to users
Made the Auth service keep track of the currently logged in user at all times. Components that need information about the currently logged in user were updated to use this angular service instead of the firebase service.

Logged in user service maintains a copy of the user with all relevant data.

in the bulk upload component, fixed the displayed name so it shows the logged in user's actual name, rather than their email

Cleaned up the validation api endpoint, made it include the information for both the User and Observation entities that are relevant to an application user.

Made the navigation bar only display links when a user with appropriate permissions is logged in.