Fix Bug: Refreshing the application logs the user out, opening a window in a new tab doesn't open a window in the current session.
Made the authService maintain the value of current user not just in its own local variables but also in local storage, which persists after refresh or opening a link in a new tab.

in the root angular component, (app.component), removed dead code. The reference to the AuthService is still necessary because it causes the auth service (which needs to check local storage for a user) to exist at the very beginning of the application (for when we refresh or open a link in a new tab)